P1-Data-Importer/rag/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
SQLAlchemy==2.0.32
pymysql==1.1.1
chromadb==0.5.5
pypdf==4.3.1
python-docx==1.1.2
